msdt-4/laboratory_works_ECM/requirements.txt
alembic==1.14.0
blinker==1.9.0
click==8.1.7
DateTime==5.5
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.7
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2024.2
setuptools==75.6.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
uWSGI==2.0.28
Werkzeug==3.1.3
XlsxWriter==3.2.0
zope.interface==7.2
